Improve cipherUpdate
1) Don't use unsafe import for functions which pass control to C for unbounded
amount of time since then a Haskell thread that runs them can't be preempted. In
fact, it might be best to import everything as safe and change some functions to
unsafe later as an optimization pass (if it's even needed).

2) Don't make a copy of the input text - when performing operations on large
bytestrings, duplicating the memory used by them is not desirable. This also
applies to other relevant functions that use asBytesLen whereas unsafeAsBytesLen
is sufficient. ByteStringS are pinned, so it's fine not to make a copy.