Add a GHA workflow that checks whitespace in Haskell files
In Haskell files, ideally trailing whitespace should be removed. We will use the
`fix-whitespace` executable to ensure this is the case (which also ensures there
are terminal newlines). We add a `fix-whitespace.sh` script that can be used to
either format files to fix the whitespace, or to check that the whitespace is
correct. We add a GHA workflow that runs the script in check mode.