`botan`: enable GHC warnings, use `GHC2021`, add export lists
When fixing the warnings, we've tried for the most part not to change the
definitions and implementations of types, classes and functions. The exported
entities should still be the same. With respect to unused definitions, we have
mostly removed them. Even if some unused definitions could have been integrated
into existing code or exported directly, we did not do such investigations for
the sake of saving time.