`SRP6`: review, refactor, fix issue #28
This includes only the relevant modules from `botan-bindings` and `botan-low`.

In particular, we work around a bug in `srp6ServerSessionStep1` in `botan-low`
where an exception is thrown from C code by `botan_srp6_server_session_step1`
because it is called twice by `srp6ServerSessionStep1`. This is bug in the Botan
C++ library, see https://github.com/randombit/botan/issues/5112.

The best we can do for now is to try to not trigger the exception. The
work-around we introduce is to make sure we call the erroring function
(`botan_srp6_server_session_step1`) only once. Previously, we were calling it
once to "query" the size of output buffers, and then again with the correct
sizes of output buffers. We can instead use
`botan_srp6_group_size`/`srp6GroupSize` to determine the correct sizes of output
buffers immediately. This is now done this way automatically where appropriate
in `botan-low`, but not in `botan-bindings`. Moreover, users can still call
`botan_srp6_server_session_step1`/`srp6ServerSessionStep1` twice if they want
to, meaning that the exception can still be triggered. To warn against this, we
make sure to include appropriate warnings in the Haddock documentation.

The other smaller changes do not affect the functionality of the code, only
documentation and such.