Review, refactor, and fix `SRP6` modules.
This includes only the relevant modules from `botan-bindings` and `botan-low`.
In particular, we work around a bug in `srp6ServerSessionStep1` (described
below). The other main changes do not affect the functionality of the code. The
existing tests are expanded upon.

So, we work around a bug in `botan-low` where an exception is thrown from C code
by `botan_srp6_server_session_step1` (`foo` for short) because it is called
twice by `srp6ServerSessionStep1` (`bar` for short). If calling `foo` twice from
`bar` *should* throw an exception, then it is not documented in the Botan C++
library why that is the case. So, the best we can do for now is to try to not
trigger the exception. The work-around we introduce is to make sure we call the
`foo` only once in `bar`. Previously, we were calling `foo` once to "query" the
size of output buffers, and then again with the correct sizes of output buffers.
We can instead use `botan_srp6_group_size`/`srp6GroupSize` to determine the
correct sizes of output buffers immediately. This is now done this way
automatically where appropriate in `botan-low`, but not in `botan-bindings`.
Moreover, users can still call
`botan_srp6_server_session_step1`/`srp6ServerSessionStep1` twice if they want
to, meaning that the exception can still be triggered. To warn against this, we
make sure to include appropriate warnings in the Haddock documentation.