`botan-low`: remove the TODO about crashes on the `keyAgreement` function.
The tests do not show any crashes in this functions, and there does not seem to
be a leak. I think the TODO was mainly there just to warn that there *had* been
a bug in the past, but it is now fixed.

I've also reformatted the code in the `KeyAgreement` module a bit